requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.4
    # via django
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
django-js-asset==1.2.2
    # via django-mptt
django-mptt==0.12.0
    # via -r requirements.in
django-summernote==0.8.11.6
    # via -r requirements.in
django-taggit-serializer==0.1.7
    # via -r requirements.in
django-taggit==1.4.0
    # via
    #   -r requirements.in
    #   django-taggit-serializer
django==3.2.3
    # via
    #   -r requirements.in
    #   django-mptt
    #   django-summernote
    #   django-taggit
    #   djangorestframework
    #   easy-thumbnails
djangorestframework==3.12.4
    # via -r requirements.in
easy-thumbnails==2.7.1
    # via -r requirements.in
factory-boy==3.2.0
    # via -r requirements.in
faker==8.2.1
    # via
    #   -r requirements.in
    #   factory-boy
idna==2.10
    # via requests
pillow==8.2.0
    # via
    #   -r requirements.in
    #   easy-thumbnails
python-dateutil==2.8.1
    # via
    #   -r requirements.in
    #   faker
python-decouple==3.4
    # via -r requirements.in
pytz==2021.1
    # via django
requests==2.25.1
    # via -r requirements.in
six==1.16.0
    # via
    #   django-taggit-serializer
    #   python-dateutil
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
urllib3==1.26.4
    # via requests
